Adds unit tests

+++ EDU.Application/UseCases/Groups/RemoveStudentUseCase.cs
@@ -23,10 +23,12 @@ namespace EDU.Application.UseCases.Groups
 
         public async Task Execute(RemoveStudentInput input)
         {
+            if (input == null) { outputPort.WriteError(""); return; }
+
             if (
                await groupRepository.Get(input.GroupId) == null ||
                await userRepository.Get(input.StudentId) == null
-              ) { outputPort.NotFound(""); }
+              ) { outputPort.NotFound(""); return; }
 
             bool success = await groupRepository.RemoveStudent(input.GroupId, input.StudentId);
             outputPort.Standart(new RemoveStudentOutput(success));
